+++ packages/react-ui/js/core/layout.ts
@@ -34,7 +34,7 @@ import {
 import {propsEqual} from '../helpers/props-helpers.js';
 import {roundedRectangle, roundedRectangleOutline} from '../rounded-rectangle-mesh.js';
 import {nineSlice} from '../nine-slice.js';
-import {NodeWrapper, Context} from './core.js';
+import {NodeWrapper, Context} from './yoga.js';
 import {DEFAULT_FONT_SIZE, TEXT_BASE_SIZE} from '../constants.js';
 
 /**
