+++ packages/react-ui/test/unit/yoga.test.ts
@@ -1,5 +1,5 @@
 import {describe, it, expect, beforeAll, vi} from 'vitest';
-import {yoga, setYoga, Context, NodeWrapper, destroyTreeForNode} from '../../js/core/core';
+import {yoga, setYoga, Context, NodeWrapper, destroyTreeForNode} from '../../js/core/yoga';
 
 describe('core module', () => {
     // Prepare a fake Yoga instance that exposes Config.create() used by Context
